refactored QuickListComponent
